requirements.txt
matplotlib
numpy
pandas
opencv-contrib-python
tensorflow-cpu
matplotlib.pyplot